+++ modules/mpi/ospray/MPIOffloadWorker.cpp
@@ -16,8 +16,9 @@
 #include "common/MPIBcastFabric.h"
 #include "common/MPICommon.h"
 #include "common/OSPWork.h"
+#include "common/Profiling.h"
 #include "common/SocketBcastFabric.h"
-#include "ospcommon/utility/getEnvVar.h"
+#include "rkcommon/utility/getEnvVar.h"
 
 #ifndef HOST_NAME_MAX
 #define HOST_NAME_MAX 4096
@@ -55,6 +56,15 @@ void runWorker(bool useMPIFabric)
       << "#w: running MPI worker process " << workerRank() << "/"
       << workerSize() << " on pid " << getpid() << "@" << hostname;
 
+#ifdef ENABLE_PROFILING
+  const std::string worker_log_file =
+      std::string(hostname) + "_" + std::to_string(workerRank()) + ".txt";
+  std::ofstream fout(worker_log_file.c_str());
+  fout << "Worker rank " << workerRank() << " on '" << hostname << "'\n"
+       << "/proc/self/status:\n"
+       << getProcStatus() << "\n=====\n";
+#endif
+
   std::unique_ptr<networking::Fabric> fabric;
   if (useMPIFabric)
     fabric = make_unique<MPIFabric>(mpicommon::world, 0);
@@ -70,6 +80,9 @@ void runWorker(bool useMPIFabric)
   std::shared_ptr<utility::OwnedArray<uint8_t>> recvBuffer =
       std::make_shared<utility::OwnedArray<uint8_t>>();
 
+#ifdef ENABLE_PROFILING
+  ProfilingPoint workerStart;
+#endif
   while (1) {
     fabric->recvBcast(cmdView);
 
@@ -83,6 +96,19 @@ void runWorker(bool useMPIFabric)
     postStatusMsg(OSPRAY_MPI_VERBOSE_LEVEL)
         << "#osp.mpi.worker: processing work " << workTag << ": "
         << work::tagName(workTag);
+
+    // We're exiting so sync out our debug log info
+#ifdef ENABLE_PROFILING
+    if (workTag == work::FINALIZE) {
+      ProfilingPoint workerEnd;
+      fout << "Worker exiting, /proc/self/status:\n"
+           << getProcStatus() << "\n======\n"
+           << "Avg. CPU % during run: "
+           << cpuUtilization(workerStart, workerEnd) << "%\n"
+           << std::flush;
+      fout.close();
+    }
+#endif
 
     dispatchWork(workTag, ospState, reader, *fabric);