+++ modules/mpi/ospray/render/WriteMultipleTileOperation.cpp
@@ -3,10 +3,10 @@
 
 #include "WriteMultipleTileOperation.h"
 #include "../fb/DistributedFrameBuffer.h"
-#include "DistributedFrameBuffer_ispc.h"
+#include "fb/DistributedFrameBuffer_ispc.h"
 
 namespace ospray {
-using namespace ospcommon;
+using namespace rkcommon;
 
 /* LiveTileOperation for image-parallel rendering, where the same image tile
  * could optionally be rendered by multiple ranks.